# notification_service/docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Notification Service API
  version: 1.0.0
  description: REST API for sending and retrieving notifications

servers:
  - url: http://localhost:3000

paths:
  /notifications:
    post:
      summary: Queue a notification for delivery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationRequest'
      responses:
        '202':
          description: Notification queued successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationQueued'
        '400':
          description: Invalid request

  /users/{userId}/notifications:
    get:
      summary: Get notifications for a user
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: skip
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
        '404':
          description: User not found

components:
  schemas:
    NotificationRequest:
      type: object
      required:
        - userId
        - channel
        - message
      properties:
        userId:
          type: string
        channel:
          type: string
          enum: [email, sms, in-app]
        message:
          type: string
        subject:
          type: string
        html:
          type: string

    NotificationQueued:
      type: object
      properties:
        trackingId:
          type: string
        statusUrl:
          type: string
        message:
          type: string

    Notification:
      type: object
      properties:
        userId:
          type: string
        channel:
          type: string
        message:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time